+++ src/astrameter/powermeter/__init__.py
@@ -12,6 +12,7 @@ from .iobroker import IoBroker
 from .json_http import JsonHttpPowermeter
 from .modbus import ModbusPowermeter
 from .mqtt import MqttPowermeter
+from .refoss import Refoss
 from .script import Script
 from .shelly import Shelly, Shelly1PM, Shelly3EM, Shelly3EMPro, ShellyEM, ShellyPlus1PM
 from .shrdzm import Shrdzm
@@ -50,6 +51,7 @@ __all__ = [
     "PidPowermeter",
     "Powermeter",
     "PowermeterWrapper",
+    "Refoss",
     "Script",
     "Shelly",
     "Shelly1PM",